requirements.txt
fastapi
uvicorn
cohere
pinecone-client
pymongo
